backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
openai==2.31.0
pydantic==2.13.0
pydantic_core==2.46.0
python-dotenv==1.2.1
sniffio==1.3.1
starlette==0.49.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
